+++ arch/xtensa/src/common/espressif/Kconfig
@@ -23,6 +23,11 @@ config ESPRESSIF_TEMP
 	---help---
 		A built-in sensor used to measure the chip's internal temperature.
 
+config ESPRESSIF_SPIFLASH
+	bool "SPI Flash"
+	depends on ARCH_CHIP_ESP32S2
+	default n
+
 menu "Internal Temperature Sensor Configuration"
 	depends on ESPRESSIF_TEMP
 
@@ -31,3 +36,120 @@ config ESPRESSIF_TEMP_PATH
 	default "dev/temp"
 
 endmenu # ESPRESSIF_TEMP
+
+config ESPRESSIF_HAVE_OTA_PARTITION
+	bool
+	default n
+
+menu "SPI Flash Configuration"
+
+choice ESPRESSIF_FLASH_MODE
+	prompt "SPI Flash mode"
+	default ESPRESSIF_FLASH_MODE_DIO
+	---help---
+		These options control how many I/O pins are used for communication
+		with the attached SPI Flash chip.
+		The option selected here is then used by esptool when flashing.
+
+config ESPRESSIF_FLASH_MODE_DIO
+	bool "Dual IO (DIO)"
+
+config ESPRESSIF_FLASH_MODE_DOUT
+	bool "Dual Output (DOUT)"
+
+config ESPRESSIF_FLASH_MODE_QIO
+	bool "Quad IO (QIO)"
+
+config ESPRESSIF_FLASH_MODE_QOUT
+	bool "Quad Output (QOUT)"
+
+endchoice # ESPRESSIF_FLASH_MODE
+
+choice ESPRESSIF_FLASH_FREQ
+	prompt "SPI Flash frequency"
+	default ESPRESSIF_FLASH_FREQ_40M
+	---help---
+		SPI Flash frequency.
+
+config ESPRESSIF_FLASH_FREQ_80M
+	bool "80 MHz"
+
+config ESPRESSIF_FLASH_FREQ_40M
+	bool "40 MHz"
+
+config ESPRESSIF_FLASH_FREQ_26M
+	bool "26 MHz"
+
+config ESPRESSIF_FLASH_FREQ_20M
+	bool "20 MHz"
+
+endchoice # ESPRESSIF_FLASH_FREQ
+
+if ESPRESSIF_HAVE_OTA_PARTITION
+
+comment "Application Image OTA Update support"
+
+config ESPRESSIF_OTA_PRIMARY_SLOT_OFFSET
+	hex "Application image primary slot offset"
+	default 0x10000
+
+config ESPRESSIF_OTA_PRIMARY_SLOT_DEVPATH
+	string "Application image primary slot device path"
+	default "/dev/ota0"
+
+config ESPRESSIF_OTA_SECONDARY_SLOT_OFFSET
+	hex "Application image secondary slot offset"
+	default 0x110000
+
+config ESPRESSIF_OTA_SECONDARY_SLOT_DEVPATH
+	string "Application image secondary slot device path"
+	default "/dev/ota1"
+
+config ESPRESSIF_OTA_SLOT_SIZE
+	hex "Application image slot size (in bytes)"
+	default 0x100000
+
+config ESPRESSIF_OTA_SCRATCH_OFFSET
+	hex "Scratch partition offset"
+	default 0x210000
+
+config ESPRESSIF_OTA_SCRATCH_SIZE
+	hex "Scratch partition size"
+	default 0x40000
+
+config ESPRESSIF_OTA_SCRATCH_DEVPATH
+	string "Scratch partition device path"
+	default "/dev/otascratch"
+
+endif
+
+if ESPRESSIF_SPIFLASH
+
+comment "General storage MTD configuration"
+
+config ESPRESSIF_MTD
+	bool "MTD driver"
+	default y
+	select MTD
+	select MTD_BYTE_WRITE
+	select MTD_PARTITION
+	---help---
+		Initialize an MTD driver for the ESP32-S2 SPI Flash, which will
+		add an entry at /dev for application access from userspace.
+
+config ESPRESSIF_SPIFLASH_MTD_BLKSIZE
+	int "Storage MTD block size"
+	default 64
+	depends on ESPRESSIF_MTD
+
+config ESPRESSIF_STORAGE_MTD_DEBUG
+	bool "Storage MTD Debug"
+	default n
+	depends on ESPRESSIF_MTD && DEBUG_FS_INFO
+	---help---
+		If this option is enabled, Storage MTD driver read and write functions
+		will output input parameters and return values (if applicable).
+
+endif # ESPRESSIF_SPIFLASH
+
+endmenu # SPI Flash Configuration
