+++ arch/xtensa/src/esp32s2/esp32s2_start.c
@@ -69,7 +69,7 @@
 #if defined(CONFIG_ESP32S2_APP_FORMAT_MCUBOOT) || \
     defined (CONFIG_ESPRESSIF_SIMPLE_BOOT)
 #  ifdef CONFIG_ESP32S2_APP_FORMAT_MCUBOOT
-#    define PRIMARY_SLOT_OFFSET   CONFIG_ESP32S2_OTA_PRIMARY_SLOT_OFFSET
+#    define PRIMARY_SLOT_OFFSET   CONFIG_ESPRESSIF_OTA_PRIMARY_SLOT_OFFSET
 #  else
      /* Force offset to the beginning of the whole image */
