+++ boards/xtensa/esp32s2/common/Kconfig
@@ -15,7 +15,7 @@ config ESP32S2_MERGE_BINS
 choice ESP32S2_SPIFLASH_FS
 	prompt "Mount SPI Flash MTD on bring-up"
 	default ESP32S2_SPIFLASH_SMARTFS
-	depends on ESP32S2_SPIFLASH
+	depends on ESPRESSIF_SPIFLASH
 	optional
 	---help---
 		Mount the SPI Flash MTD with the selected File System format on board
