+++ boards/xtensa/esp32s2/common/src/Make.defs
@@ -36,7 +36,7 @@ ifeq ($(CONFIG_SPI_SLAVE_DRIVER),y)
   CSRCS += esp32s2_board_spislavedev.c
 endif
 
-ifeq ($(CONFIG_ESP32S2_SPIFLASH),y)
+ifeq ($(CONFIG_ESPRESSIF_SPIFLASH),y)
   CSRCS += esp32s2_board_spiflash.c
 endif
