+++ boards/xtensa/esp32s2/esp32s2-saola-1/configs/mcuboot_nsh/defconfig
@@ -20,7 +20,6 @@ CONFIG_ARCH_XTENSA=y
 CONFIG_BOARD_LOOPSPERMSEC=16717
 CONFIG_BUILTIN=y
 CONFIG_ESP32S2_APP_FORMAT_MCUBOOT=y
-CONFIG_ESP32S2_SPIFLASH=y
 CONFIG_ESP32S2_UART0=y
 CONFIG_FS_PROCFS=y
 CONFIG_HAVE_CXX=y
