+++ boards/xtensa/esp32s2/esp32s2-saola-1/configs/spiflash/defconfig
@@ -19,17 +19,9 @@ CONFIG_ARCH_STACKDUMP=y
 CONFIG_ARCH_XTENSA=y
 CONFIG_BOARD_LOOPSPERMSEC=16717
 CONFIG_BUILTIN=y
-CONFIG_DEBUG_ASSERTIONS=y
-CONFIG_DEBUG_ASSERTIONS_EXPRESSION=y
-CONFIG_DEBUG_FEATURES=y
-CONFIG_DEBUG_FS=y
-CONFIG_DEBUG_FS_ERROR=y
-CONFIG_DEBUG_FS_INFO=y
-CONFIG_DEBUG_FS_WARN=y
-CONFIG_DEBUG_FULLOPT=y
-CONFIG_DEBUG_SYMBOLS=y
-CONFIG_ESP32S2_SPIFLASH=y
+CONFIG_ESP32S2_SPIFLASH_SMARTFS=y
 CONFIG_ESP32S2_UART0=y
+CONFIG_ESPRESSIF_SPIFLASH=y
 CONFIG_FS_PROCFS=y
 CONFIG_HAVE_CXX=y
 CONFIG_HAVE_CXXINITIALIZE=y
@@ -37,7 +29,6 @@ CONFIG_IDLETHREAD_STACKSIZE=3072
 CONFIG_INIT_ENTRYPOINT="nsh_main"
 CONFIG_INIT_STACKSIZE=3072
 CONFIG_INTELHEX_BINARY=y
-CONFIG_NDEBUG=y
 CONFIG_NSH_ARCHINIT=y
 CONFIG_NSH_BUILTIN_APPS=y
 CONFIG_NSH_FILEIOSIZE=512
