+++ tools/esp32s2/Config.mk
@@ -32,23 +32,23 @@ else ifeq ($(CONFIG_ESP32S2_FLASH_16M),y)
 	FLASH_SIZE := 16MB
 endif
 
-ifeq ($(CONFIG_ESP32S2_FLASH_MODE_DIO),y)
+ifeq ($(CONFIG_ESPRESSIF_FLASH_MODE_DIO),y)
 	FLASH_MODE := dio
-else ifeq ($(CONFIG_ESP32S2_FLASH_MODE_DOUT),y)
+else ifeq ($(CONFIG_ESPRESSIF_FLASH_MODE_DOUT),y)
 	FLASH_MODE := dout
-else ifeq ($(CONFIG_ESP32S2_FLASH_MODE_QIO),y)
+else ifeq ($(CONFIG_ESPRESSIF_FLASH_MODE_QIO),y)
 	FLASH_MODE := qio
-else ifeq ($(CONFIG_ESP32S2_FLASH_MODE_QOUT),y)
+else ifeq ($(CONFIG_ESPRESSIF_FLASH_MODE_QOUT),y)
 	FLASH_MODE := qout
 endif
 
-ifeq ($(CONFIG_ESP32S2_FLASH_FREQ_80M),y)
+ifeq ($(CONFIG_ESPRESSIF_FLASH_FREQ_80M),y)
 	FLASH_FREQ := 80m
-else ifeq ($(CONFIG_ESP32S2_FLASH_FREQ_40M),y)
+else ifeq ($(CONFIG_ESPRESSIF_FLASH_FREQ_40M),y)
 	FLASH_FREQ := 40m
-else ifeq ($(CONFIG_ESP32S2_FLASH_FREQ_26M),y)
+else ifeq ($(CONFIG_ESPRESSIF_FLASH_FREQ_26M),y)
 	FLASH_FREQ := 26m
-else ifeq ($(CONFIG_ESP32S2_FLASH_FREQ_20M),y)
+else ifeq ($(CONFIG_ESPRESSIF_FLASH_FREQ_20M),y)
 	FLASH_FREQ := 20m
 endif
 
@@ -89,10 +89,10 @@ endif
 ifeq ($(CONFIG_ESP32S2_APP_FORMAT_MCUBOOT),y)
 	ifeq ($(CONFIG_ESP32S2_ESPTOOL_TARGET_PRIMARY),y)
 		VERIFIED   := --confirm
-		APP_OFFSET := $(CONFIG_ESP32S2_OTA_PRIMARY_SLOT_OFFSET)
+		APP_OFFSET := $(CONFIG_ESPRESSIF_OTA_PRIMARY_SLOT_OFFSET)
 	else ifeq ($(CONFIG_ESP32S2_ESPTOOL_TARGET_SECONDARY),y)
 		VERIFIED   :=
-		APP_OFFSET := $(CONFIG_ESP32S2_OTA_SECONDARY_SLOT_OFFSET)
+		APP_OFFSET := $(CONFIG_ESPRESSIF_OTA_SECONDARY_SLOT_OFFSET)
 	endif
 
 	ifeq ($(CONFIG_ESP32S2_SECURE_BOOT),y)
@@ -111,7 +111,7 @@ ifeq ($(CONFIG_ESP32S2_APP_FORMAT_MCUBOOT),y)
 
 	IMGTOOL_SIGN_ARGS := --pad $(VERIFIED) $(IMGTOOL_ALIGN_ARGS) -v 0 -s auto \
 		-H $(CONFIG_ESP32S2_APP_MCUBOOT_HEADER_SIZE) --pad-header \
-		-S $(CONFIG_ESP32S2_OTA_SLOT_SIZE)
+		-S $(CONFIG_ESPRESSIF_OTA_SLOT_SIZE)
 else
 #   CONFIG_ESPRESSIF_SIMPLE_BOOT
